refactor: move checkpoint simulation overrides builder to stdlib
buildCheckpointSimulationOverridesPlan and computePipelinedParentFeeHeader
are consumed by both the sequencer and the node's public-calls simulator, so
they move from sequencer-client to @aztec/stdlib/checkpoint. Drops the
chain-state-overrides sub-export that existed only to share them.
